requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Azure Cosmos DB
azure-cosmos>=4.5.1

# Azure OpenAI
openai>=1.3.5

# Environment variables
python-dotenv>=1.0.0

# Date handling
python-dateutil>=2.8.2

# Testing
pytest>=7.4.3
pytest-django>=4.7.0

# Validation
pydantic>=2.0
